types file added and filter array state and related functions moved to results page
