+++ components/Map.tsx
@@ -6,16 +6,7 @@ import { Popup } from 'react-leaflet';
 import * as bardata from '../bar-data.json';
 import { useEffect, useState } from 'react';
 import styles from '../styles/map.module.css'
-
-type singleBarType = {
-    name: string, 
-    address: string,
-    postcode: string, 
-    cost: number,
-    rating: number, 
-    coordinates: [number,number],
-    website: string
-}
+import { singleBarType } from '../data/types';
 
 
 export default function Map(){
